requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Security-Too==5.3.2
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0